Improve tracking of chatting users
Users are recorded in the DB as chatting when the chat history is
loaded.  Previously, this was only done if chat had changed since
their last view.  This meant that the chat count would go to zero
if nobody was actively chatting even if people were viewing it.

Now users are ALWAYS recorded as chatting when the history is
loaded, so the chat count should reflect the number of people
viewing at all times.

This puts slightly more load on the database (that is, one
additional write per chatting user per five seconds or whatever).