Fix active list
7989d3d's approach to adding chat users to the active list was
broken (array_merge() with numeric keys doesn't merge items
with the same key).  This probably works better.

As a sidenote, this increases CPU usage slightly as it requires
resorting (of mostly sorted items).  The alternative would be to
change the active members database query to include chatters
from the outset.